+++ hutool-poi/src/main/java/cn/hutool/poi/excel/ExcelBase.java
@@ -145,7 +145,7 @@ public class ExcelBase<T extends ExcelBase<T>> implements Closeable {
 	}
 
 	/**
-	 * 获取指定坐标单元格，单元格不存在时返回<code>null</code>
+	 * 获取指定坐标单元格，单元格不存在时返回{@code null}
 	 *
 	 * @param locationRef 单元格地址标识符，例如A11，B5
 	 * @return {@link Cell}
@@ -157,7 +157,7 @@ public class ExcelBase<T extends ExcelBase<T>> implements Closeable {
 	}
 
 	/**
-	 * 获取指定坐标单元格，单元格不存在时返回<code>null</code>
+	 * 获取指定坐标单元格，单元格不存在时返回{@code null}
 	 *
 	 * @param x X坐标，从0计数，即列号
 	 * @param y Y坐标，从0计数，即行号
@@ -193,7 +193,7 @@ public class ExcelBase<T extends ExcelBase<T>> implements Closeable {
 	}
 
 	/**
-	 * 获取指定坐标单元格，如果isCreateIfNotExist为false，则在单元格不存在时返回<code>null</code>
+	 * 获取指定坐标单元格，如果isCreateIfNotExist为false，则在单元格不存在时返回{@code null}
 	 *
 	 * @param locationRef        单元格地址标识符，例如A11，B5
 	 * @param isCreateIfNotExist 单元格不存在时是否创建
@@ -206,7 +206,7 @@ public class ExcelBase<T extends ExcelBase<T>> implements Closeable {
 	}
 
 	/**
-	 * 获取指定坐标单元格，如果isCreateIfNotExist为false，则在单元格不存在时返回<code>null</code>
+	 * 获取指定坐标单元格，如果isCreateIfNotExist为false，则在单元格不存在时返回{@code null}
 	 *
 	 * @param x                  X坐标，从0计数，即列号
 	 * @param y                  Y坐标，从0计数，即行号