+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### 新特性
 * 【core   】     DynaBean.create增加重载方法（pr#245@Gitee）
 * 【core   】     IdcardUtil增加重载是否忽略大小写（issue#1348@Github）
+* 【poi    】     SheetRidReader增加getRidByIndex方法（issue#1342@Github）
 
 ### Bug修复